+++ yggdrasil_decision_forests/learner/random_forest/random_forest.h
@@ -93,6 +93,7 @@ class RandomForestLearner : public AbstractLearner {
   model::proto::LearnerCapabilities Capabilities() const override {
     model::proto::LearnerCapabilities capabilities;
     capabilities.set_support_max_training_duration(true);
+    capabilities.set_support_max_model_size_in_memory(true);
     return capabilities;
   }
 };